refactor: exclude vod_id and id from VOD relationships, change Game timestamps to Int
- Update VodResponse interface to explicitly select fields for vod_uploads, chapters, and games
- Exclude vod_id and id from relationship responses in getVods and getVodById
- Change Game.start_time and Game.end_time from Decimal to Int in schema
- Update migration SQL to use INTEGER instead of DECIMAL for Game timestamps
- Convert Decimal to Int in migration script when migrating legacy Game data
- Update VodResponse.games interface to use number instead of string for timestamps

+++ prisma/migrations/20240101000000_add_normalized_schema/migration.sql
@@ -45,8 +45,8 @@ CREATE TABLE "emotes" (
 CREATE TABLE "games" (
     "id" SERIAL NOT NULL,
     "vod_id" TEXT NOT NULL,
-    "start_time" DECIMAL(65,30),
-    "end_time" DECIMAL(65,30),
+    "start_time" INTEGER,
+    "end_time" INTEGER,
     "video_provider" TEXT,
     "video_id" TEXT,
     "thumbnail_url" TEXT,
